Move reconciliation columns file
