+++ tests/unit/test_reconciliation.py
@@ -17,7 +17,7 @@ from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
 from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
     CqcLocationCleanedColumns as CQCLClean,
 )
-from utils.reconciliation_utils.reconciliation_values import (
+from utils.column_names.reconciliation_values import (
     ReconciliationColumns as ReconColumn,
 )
