+++ utils/column_values/categorical_column_values.py
@@ -77,7 +77,9 @@ class Services(ColumnValues):
     hospice_services: str = "Hospice services"
     domiciliary_care_service: str = "Domiciliary care service"
     remote_clinical_advice_service: str = "Remote clinical advice service"
-    acute_services_without_overnight_beds: str = "Acute services without overnight beds / listed acute services with or without overnight beds"
+    acute_services_without_overnight_beds: str = (
+        "Acute services without overnight beds / listed acute services with or without overnight beds"
+    )
     specialist_college_service: str = "Specialist college service"
     ambulance_service: str = "Ambulance service"
     extra_care_housing_services: str = "Extra Care housing services"
@@ -107,7 +109,9 @@ class Services(ColumnValues):
     rehabilitation_services: str = "Rehabilitation services"
     doctors_treatment_service: str = "Doctors treatment service"
     hospice_services_at_home: str = "Hospice services at home"
-    hospital_services_for_people_with_mental_health_needs: str = "Hospital services for people with mental health needs, learning disabilities and problems with substance misuse"
+    hospital_services_for_people_with_mental_health_needs: str = (
+        "Hospital services for people with mental health needs, learning disabilities and problems with substance misuse"
+    )
 
 
 @dataclass
@@ -440,3 +444,15 @@ class CQCRatingsValues(ColumnValues):
     historic: str = "Historic"
     outstanding: str = "Outstanding"
     good: str = "Good"
+
+
+@dataclass
+class ReconciliationValues(ColumnValues):
+    singles_and_subs: str = "singles_and_subs"
+    parents: str = "parents"
+    single_sub_subject_value: str = "CQC Reconcilliation Work"
+    parent_subject_value: str = "CQC Reconcilliation Work - Parent"
+    single_sub_deregistered_description: str = "Potential (new): Deregistered ID"
+    single_sub_reg_type_description: str = "Potential (new): Regtype"
+    is_parent: str = "Yes"
+    is_not_parent: str = "No"
